Canvas; add canvas.scale property (see long desc)
This value is to help when rendering the canvas at a scaled size, like
on a high DPI screen. As the doc line says, this value is only a hint
for the rendering, no internal logic, coords, or values are affected.
This keeps all coordinates consistent and relative, and only the
resulting use of those coordinates needs to be adjusted. This is true
for things like rendering and mouse inputs, which was always true (when
dealing with cameras).