render; luxe; Image; NOTE; changed to be deterministic, see long desc
The image control previously tried to load the image path but this is a
problem for a variety of reasons.

For one, all the state changes made to the control either directly or
by the controls propagating them are ignored until the image loads.

The second is that the control can be destroyed before the image loads.

The third is that the image can fail to load.

The fourth is that creating the control based on the image size isn't
possible in this model either.

In practice, this was a left over from before luxe resources were
treated deterministically, and now this falls in line with the rest of
the engine. You can't use a resource until it is loaded, if you create
the control with an image that isn't loaded the control will just
display a white box for now.

For the above issues, giving you the control over failed images, timing
of creating and destroying, sizes are all in your control rightfully
now, and, the control can now be deterministic about it's state so it's
easier to reason about and deal with.

You can use :

Luxe.resources.load_texture(path).then(function(_texture:phoenix.Texture
) {
   new mint.Image({
      w:_texture.width, h: _texture.height,
      ...
   })

   -- OR --

   (cast image.renderer : mint.render.luxe.Image).visual.texture =
_texture;

});

Whichever works best for your use case.